ECRv2 greatly refactored version of ECR
In the aim of refactoring the ECR code, a new implementation is provided to interface
with the kernel algorithm. Currently, it depends on LLVM for cli parsing only but now,
different LLVM libs can be used easily.

Cuda file is kept to minimum for easier maintainance. This
implementation is path and parameter independent except the batch size
limited to 1. The aim is to provide different tools based on the same
Cuda source.